backend/backend/requirements.txt
asgiref==3.11.0
certifi==2026.2.25
charset-normalizer==3.4.6
Django==6.0.2
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
et_xmlfile==2.0.0
geopandas==1.1.3
idna==3.11
networkx==3.6.1
numpy==2.4.0
openpyxl==3.1.5
osmnx==2.1.0
packaging==26.0
pandas==2.3.3
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyogrio==0.12.1
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
